client: Fixed log output for amd64 platform

+++ bindings/cpp/callback_p.h
@@ -394,7 +394,8 @@ class multigroup_callback
 		{
 			dnet_log_raw(sess.get_native_node(),
 				DNET_LOG_DEBUG, "%s: multigroup_callback::handle: cmd: %s, trans: %llx, status: %d, flags: %llx, group: %d: %zd/%zd, priv: %p\n",
-					dnet_dump_id(&cmd->id), dnet_cmd_string(cmd->cmd), cmd->trans, cmd->status, cmd->flags,
+					dnet_dump_id(&cmd->id), dnet_cmd_string(cmd->cmd), static_cast<long long unsigned>(cmd->trans),
+					cmd->status, static_cast<long long unsigned>(cmd->flags),
 					m_group_index < groups.size() ? groups[m_group_index] : -1, m_group_index, groups.size(), priv);
 
 			if (cb.handle(state, cmd, func, priv)) {
